+++ a4md/common/ims_reader.cxx
@@ -5,7 +5,7 @@ std::vector<Chunk*> IMSReader::get_chunks(int num_chunks)
     throw NotImplementedException("IMSReader::get_chunks should not be called. It should be overridden in a concrete class\n");
 }
 
-std::vector<Chunk*> IMSReader::get_chunks(int chunks_from, int chunks_to)
+std::vector<Chunk*> IMSReader::get_chunks(unsigned long int chunks_from, unsigned long int chunks_to)
 {
     throw NotImplementedException("IMSReader::get_chunks should not be called. It should be overridden in a concrete class\n");
 }
