+++ a4md/common/ims_reader.h
@@ -7,6 +7,6 @@ class IMSReader : public Chunker
 {
     public:
         std::vector<Chunk*> get_chunks(int num_chunks) override;
-        std::vector<Chunk*> get_chunks(int chunk_id_from, int chunk_id_to) override;
+        std::vector<Chunk*> get_chunks(unsigned long int chunk_id_from, unsigned long int chunk_id_to) override;
 };
 #endif
